requirements.txt
asgiref==3.8.1
Django==5.0.4
psycopg2-binary==2.9.9
sqlparse==0.4.4
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
Markdown==3.6
psycopg2-binary==2.9.9
sqlparse==0.4.4
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
Markdown==3.6
psycopg2-binary==2.9.9
sqlparse==0.4.4
asgiref==3.8.1
Django==5.0.4
django-nested-admin==4.0.2
djangorestframework==3.15.1
Markdown==3.6
psycopg2-binary==2.9.9
python-monkey-business==1.0.0
setuptools==69.5.1
six==1.16.0
sqlparse==0.4.4
